requirements.txt
streamlit
pypdf
PyPDF2
pinecone-client
openai
langchain
langchain-community
langchain-core
langchain-experimental
googletrans==4.0.0-rc1
translators
sentence-transformers